+++ knotx-knot/knotx-knot-fragment-splitter/src/main/java/io/knotx/splitter/FragmentSplitterVerticle.java
@@ -24,7 +24,7 @@ import io.vertx.core.eventbus.MessageConsumer;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
-import io.vertx.serviceproxy.ProxyHelper;
+import io.vertx.serviceproxy.ServiceBinder;
 
 public class FragmentSplitterVerticle extends AbstractVerticle {
 
@@ -33,6 +33,8 @@ public class FragmentSplitterVerticle extends AbstractVerticle {
   private FragmentSplitterConfiguration configuration;
 
   private MessageConsumer<JsonObject> consumer;
+
+  private ServiceBinder serviceBinder;
 
   @Override
   public void init(Vertx vertx, Context context) {
@@ -45,13 +47,14 @@ public class FragmentSplitterVerticle extends AbstractVerticle {
     LOGGER.info("Starting <{}>", this.getClass().getSimpleName());
 
     //register the service proxy on event bus
-    consumer = ProxyHelper
-        .registerService(KnotProxy.class, vertx, new FragmentSplitterKnotProxyImpl(),
-            configuration.getAddress());
+    serviceBinder = new ServiceBinder(getVertx());
+    consumer = serviceBinder
+        .setAddress(configuration.getAddress())
+        .register(KnotProxy.class, new FragmentSplitterKnotProxyImpl());
   }
 
   @Override
   public void stop() throws Exception {
-    ProxyHelper.unregisterService(consumer);
+    serviceBinder.unregister(consumer);
   }
 }